requirements.txt
torch==2.8.0
transformers==4.55.0
tokenizers==0.21.4
pandas==2.1.3
tqdm==4.66.1
opencv-python==4.8.0.76
numpy==1.26.2
Pillow==10.1.0
matplotlib==3.8.2
# 建议: 在您的开发环境中运行 'pip freeze > requirements.txt' 来获取精确的版本号，以确保与您当前环境一致。
